simplified build, and cleaned up files
