+++ src/tfr_dumping/src/dumping_action_server.cpp
@@ -1,5 +1,12 @@
-#include <dumping_action_server.h>
-
+#include <ros/ros.h>
+#include <geometry_msgs/Twist.h>
+#include <tfr_msgs/EmptyAction.h>
+#include <tfr_msgs/ArucoIntegrateAction.h>
+#include <tfr_msgs/WrappedImage.h>
+#include <sensor_msgs/Image.h>
+#include <image_transport/image_transport.h>
+#include <actionlib/server/simple_action_server.h>
+#include <actionlib/client/simple_action_client.h>
 /*
  * The dumping action server, it backs up the rover into the navigational aid
  * slowly.
@@ -7,7 +14,7 @@
  * Its first step is to make sure it can see the aruco board, it will abort the
  * mission if it can't. 
  *
- * I backs up at a set speed until it get's really close and loses sight of the
+ * It backs up at a set speed until it get's really close and loses sight of the
  * board. When it is blind, it drives straight back, goes slower. 
  *
  * It stops when the light detector get's triggered.
@@ -15,140 +22,178 @@
  * It requires a service where it can get the most recent image on demand for
  * the camera of interest for backing up. 
  * */
-Dumper::Dumper(ros::NodeHandle &node, const std::string &service_name,
-        const DumpingConstraints &c) :
-    server{node, "dump", boost::bind(&Dumper::dump, this, _1), false},
-    detector{"light_detection"},
-    //TODO add dumping controller here
-    //TODO add aruco here
-    constraints{c}
+class Dumper
 {
-    ROS_INFO("dumping action server initializing");
-    detector.waitForServer();
-
-    image_client = node.serviceClient<tfr_msgs::WrappedImage>(service_name);
-
-    velocity_publisher = node.advertise<geometry_msgs::Twist>("cmd_vel", 10);
-
-    server.start();
-    ROS_INFO("dumping action server initialized");
-}
-
-/*
- * The business logic of the action server.
- */
-void Dumper::dump(const tfr_msgs::EmptyGoalConstPtr &goal) 
-{  
-    //check to make sure we can see the board
-    tfr_msgs::ArucoIntegrateResult initial_estimate{};
-    getArucoEstimate(initial_estimate);
-    if (initial_estimate.markers_detected == 0)
-    {
-        server.setAborted();
-        return;
-    }
-
-    //initialize
-    tfr_msgs::EmptyGoal empty_goal{};
-    detector.sendGoal(empty_goal);
-
-    //loop until we see the light
-    while (detector.getState() != actionlib::SimpleClientGoalState::SUCCEEDED)
-    {
-        //handle preemption
-        if (server.isPreemptRequested() || !ros::ok())
+    public:        
+        struct DumpingConstraints
         {
-            stopMoving();
-            server.setPreempted();
-            return;
+            private:
+                double min_lin_vel, max_lin_vel, min_ang_vel, max_ang_vel;
+            public:
+                DumpingConstraints(double min_lin, double max_lin, 
+                        double min_ang, double max_ang):
+                    min_lin_vel(min_lin), max_lin_vel(max_lin),
+                    min_ang_vel(min_ang), max_ang_vel(max_ang) {}
+                double getMinLinVel() const {return min_lin_vel;}
+                double getMaxLinVel() const {return max_lin_vel;}
+                double getMinAngVel() const {return min_ang_vel;}
+                double getMaxAngVel() const {return max_ang_vel;}
+        };
+
+        
+        Dumper(ros::NodeHandle &node, const std::string &service_name,
+                const DumpingConstraints &c) :
+            server{node, "dump", boost::bind(&Dumper::dump, this, _1), false},
+            detector{"light_detection"},
+            //TODO add dumping controller here
+            //TODO add aruco here
+            constraints{c}
+        {
+            ROS_INFO("dumping action server initializing");
+            detector.waitForServer();
+
+            image_client = node.serviceClient<tfr_msgs::WrappedImage>(service_name);
+
+            velocity_publisher = node.advertise<geometry_msgs::Twist>("cmd_vel", 10);
+
+            server.start();
+            ROS_INFO("dumping action server initialized");
         }
 
-        //get the most recent aruco reading
-        tfr_msgs::ArucoIntegrateResult estimate{};
-        getArucoEstimate(estimate);
+        ~Dumper() = default;
+        Dumper(const Dumper&) = delete;
+        Dumper& operator=(const Dumper&) = delete;
+        Dumper(Dumper&&) = delete;
+        Dumper& operator=(Dumper&&) = delete;
 
-        //send motor commands
-        if (estimate.markers_detected == 0)
-            moveBlind();
-        else
+    private:
+        actionlib::SimpleActionServer<tfr_msgs::EmptyAction> server;
+        actionlib::SimpleActionClient<tfr_msgs::EmptyAction> detector;
+
+        //TODO add the Dumping Controller here
+        //TODO add aruco here
+
+        ros::ServiceClient image_client;
+        ros::Publisher velocity_publisher;
+
+        const DumpingConstraints &constraints; 
+
+        /*
+         * The business logic of the action server.
+         */
+        void dump(const tfr_msgs::EmptyGoalConstPtr &goal) 
+        {  
+            //check to make sure we can see the board
+            tfr_msgs::ArucoIntegrateResult initial_estimate{};
+            getArucoEstimate(initial_estimate);
+            if (initial_estimate.markers_detected == 0)
+            {
+                server.setAborted();
+                return;
+            }
+
+            //initialize
+            tfr_msgs::EmptyGoal empty_goal{};
+            detector.sendGoal(empty_goal);
+
+            //loop until we see the light
+            while (detector.getState() != actionlib::SimpleClientGoalState::SUCCEEDED)
+            {
+                //handle preemption
+                if (server.isPreemptRequested() || !ros::ok())
+                {
+                    stopMoving();
+                    server.setPreempted();
+                    return;
+                }
+
+                //get the most recent aruco reading
+                tfr_msgs::ArucoIntegrateResult estimate{};
+                getArucoEstimate(estimate);
+
+                //send motor commands
+                if (estimate.markers_detected == 0)
+                    moveBlind();
+                else
+                {
+                    geometry_msgs::Twist cmd{};
+                    updateControlMsg(estimate, cmd);
+                    velocity_publisher.publish(cmd);
+                }
+            }
+            //TODO raise the bin
+            stopMoving();
+            server.setSucceeded();
+        }
+
+        /*
+         *  Back up and turn slightly to match the orientation of the aruco board
+         * */
+        void updateControlMsg(const tfr_msgs::ArucoIntegrateResult &estimate,
+                geometry_msgs::Twist &cmd)
         {
+            //back up
+            cmd.linear.x = -1 * constraints.getMaxLinVel();
+            /*
+             * Maintenence note:
+             *
+             * How do we decide if we are going left or right?
+             * 
+             * Well the estimate will return a pose describing displacement from our
+             * rear camera, a +y displacement means the center of the board is to the
+             * left(ccw), -y to the right (cw). 
+             *
+             * This conforms to rep 103
+             * */
+            int sign = (estimate.pose.position.y < 0) ? -1 : 1;
+            cmd.angular.z = sign*constraints.getMaxAngVel();
+        }
+
+        /*
+         * back up slowwwwly we can't see
+         * */
+        void moveBlind()
+        {
+            ROS_INFO("backing up blind");
             geometry_msgs::Twist cmd{};
-            updateControlMsg(estimate, cmd);
+            cmd.linear.x = -1*constraints.getMinLinVel();
+            cmd.angular.z = 0;
             velocity_publisher.publish(cmd);
         }
-    }
-    //TODO raise the bin
-    stopMoving();
-    server.setSucceeded();
-}
 
-/*
- *  Back up and turn slightly to match the orientation of the aruco board
- * */
-void Dumper::updateControlMsg(const tfr_msgs::ArucoIntegrateResult &estimate,
-        geometry_msgs::Twist &cmd)
-{
-    //back up
-    cmd.linear.x = -1 * constraints.getMaxLinVel();
-    /*
-     * Maintenence note:
-     *
-     * How do we decide if we are going left or right?
-     * 
-     * Well the estimate will return a pose describing displacement from our
-     * rear camera, a +y displacement means the center of the board is to the
-     * left(ccw), -y to the right (cw). 
-     *
-     * This conforms to rep 103
-     * */
-    int sign = (estimate.pose.position.y < 0) ? -1 : 1;
-    cmd.angular.z = sign*constraints.getMaxAngVel();
-}
+        /*
+         *  Stop Moving 
+         * */
+        void stopMoving()
+        {
+            ROS_INFO("Stopped");
+            geometry_msgs::Twist cmd{};
+            cmd.linear.x = 0;
+            cmd.angular.z =0;
+            velocity_publisher.publish(cmd);
+        }
 
-/*
- * back up slowwwwly we can't see
- * */
-void Dumper::moveBlind()
-{
-    ROS_INFO("backing up blind");
-    geometry_msgs::Twist cmd{};
-    cmd.linear.x = -1*constraints.getMinLinVel();
-    cmd.angular.z = 0;
-    velocity_publisher.publish(cmd);
-}
-
-/*
- *  Stop Moving 
- * */
-void Dumper::stopMoving()
-{
-    ROS_INFO("Stopped");
-    geometry_msgs::Twist cmd{};
-    cmd.linear.x = 0;
-    cmd.angular.z =0;
-    velocity_publisher.publish(cmd);
-}
-
-/*
- * Gets the most recent position estimate from the aruco service
- */
-void Dumper::getArucoEstimate(tfr_msgs::ArucoIntegrateResult &result)
-{
-    tfr_msgs::WrappedImage image_request{};
-    tfr_msgs::ArucoIntegrateGoal goal{};
-    //grab the most recent image
-    image_client.call(image_request);
-    goal.image = image_request.response.image;
-    //TODO hook up aruco here
-    result.markers_detected = 1;
-    result.pose.position.x= 0.5;
-    result.pose.position.y= -0.1;
-    result.pose.position.z= -0.02;
-    result.pose.orientation.x = 0;
-    result.pose.orientation.y = 0;
-    result.pose.orientation.z = 1;
-    result.pose.orientation.w = 0;
-}
+        /*
+         * Gets the most recent position estimate from the aruco service
+         */
+        void getArucoEstimate(tfr_msgs::ArucoIntegrateResult &result)
+        {
+            tfr_msgs::WrappedImage image_request{};
+            tfr_msgs::ArucoIntegrateGoal goal{};
+            //grab the most recent image
+            image_client.call(image_request);
+            goal.image = image_request.response.image;
+            //TODO hook up aruco here
+            result.markers_detected = 1;
+            result.pose.position.x= 0.5;
+            result.pose.position.y= -0.1;
+            result.pose.position.z= -0.02;
+            result.pose.orientation.x = 0;
+            result.pose.orientation.y = 0;
+            result.pose.orientation.z = 1;
+            result.pose.orientation.w = 0;
+        }
+};
 
 int main(int argc, char **argv)
 {
